Move writing bitmap inode from save_inode() to rollup for atomic-commit
In atomic-commit case, bitmap inode is written in rollup stage. So,
this remove it from sync_inodes(). Then, all flushing codes for
atomic-commit are moved to do_commit().

With this change, bitmap inode may not be written after mkfs done
(i.e. open_inode() may fail). To handle this case, this adds basically

	inode = iget()
	if (not found inode)
		inode = create_inode()

which is called "iget_or_create_inode()".

For now, the user of iget_or_create_inode() is only bitmap inode.

Signed-off-by: OGAWA Hirofumi <hirofumi@mail.parknet.co.jp>