Fix buffer->count refcount
Now, we are using buffer->count == 0 even if buffer is hashed. So, we
lookup ->count == 0 in hash to reclaim buffer. It was no problem until
blockdirty() is introduced.

blockdirty() removes an original buffer of clone buffer from hash, so,
original buffer is not reclaimable from hash anymore (now, original
buffer is manually freeing by evict_buffer()).

So, we want to reclaim at blockput() if ->count == 0 though, hashed
buffer is also ->count == 0. The buffer->count refcount is unclear to
reclaim.

To fix it, this takes additional refcount for hash (and remove
refcount if unhashed). With this, we can reclaim buffer if ->count == 0,
and we can use blockput() simply for unhashed buffer too.

Signed-off-by: OGAWA Hirofumi <hirofumi@mail.parknet.co.jp>