Implement ileaf_write()
We are now using iattr.c and ileaf.c directly for inode. It means
inode.c is depending on ileaf/iattr format more or less.

The series try to abstract those and try to not depend on ileaf/iattr
format.

This implement ileaf_write() to use as ->leaf_write handler. This is
implemented by ileaf_attr_ops will be provided from iattr.c and
orphan.c.

Signed-off-by: OGAWA Hirofumi <hirofumi@mail.parknet.co.jp>