Move inode deletion to backend, Fix the hole extent handling
If there is hole extents on delete path (i.e. truncate and deletion
was done on same delta), i_size might be changed and i_size is not
represent last extent in dtree.

So, if we clear hole extents, then we have to check extents to free
even if i_size == 0.

Signed-off-by: OGAWA Hirofumi <hirofumi@mail.parknet.co.jp>