Fix inode LRU for reclaim after clearing dirty of inode
iput_final() doesn't add inode to LRU list if I_DIRTY.  Grab refcount
to tell inode state was changed to iput(). So memory reclaimer can't
free the clean inode.

This become the cause of out-of-memory.

To fix this, we call iget/iput() when clearing dirty of inode.

FIXME: iget() and iput() set I_REFERENCED, but we would not want to
set I_REFERENCED for clearing dirty.

Signed-off-by: OGAWA Hirofumi <hirofumi@mail.parknet.co.jp>